Keep only the single "change time" regularization algorithm
We had two algorithms for step regularization: one was for
interval_type="average" and one for interval_type="sum". The first one
was interpolating whereas the second one was merely getting a nearby
record and changing its time (details in the documentation). We decided
that the first one was too smart and have got rid of it.